% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{addPlotID}
\alias{addPlotID}
\title{Adds PlotID column to LAI data}
\usage{
addPlotID(data, path)
}
\arguments{
\item{data}{data.frame with LAI data}

\item{path}{of the original data}
}
\value{
data.frame with PlotID column
}
\description{
Information about PlotID is given via additional
tab delimited files.
}
