% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{addPlotIDFromData}
\alias{addPlotIDFromData}
\title{Adds PlotID column to LAI data}
\usage{
addPlotIDFromData(data, ids, mids)
}
\arguments{
\item{data}{data.frame with LAI data}

\item{ids}{dataframe with plotids}

\item{mids}{dataframe with measurementids}
}
\value{
data.frame with PlotID column
}
\description{
Information about PlotID is given via additional
tab delimited files.
}
