% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{convertSunScanDirectory}
\alias{convertSunScanDirectory}
\title{Converts all SunScan data files in a directory (and it's subdirectories)}
\usage{
convertSunScanDirectory(
  directory,
  prefix = "",
  extension = ".txt",
  reportscript = ""
)
}
\arguments{
\item{directory}{to search for data files}

\item{prefix}{of files to take into accoung}

\item{extension}{of files to take into account}

\item{reportscript}{R script that is used to generate a report for each converted file}
}
\description{
Converts all SunScan data files in a directory (and it's subdirectories)
}
