% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{convertSunscanFile}
\alias{convertSunscanFile}
\title{Converts a SunScan data file}
\usage{
convertSunscanFile(file, inputfolder, outputfolder)
}
\arguments{
\item{file}{name of file}

\item{inputfolder}{input folder}

\item{outputfolder}{output folder}
}
\value{
vector with filenames of converted data
}
\description{
Converts a SunScan data file
}
