% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{countMeasurements}
\alias{countMeasurements}
\title{Counts measurements}
\usage{
countMeasurements(data)
}
\arguments{
\item{data}{imported sunscan data}
}
\value{
dataframe with number of measurements
}
\description{
Counts measurements
}
