% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createBoxplot}
\alias{createBoxplot}
\title{Creates a boxplot for LAI data}
\usage{
createBoxplot(data)
}
\arguments{
\item{data}{data.frame with LAI data}
}
\value{
ggplot graph with boxplot
}
\description{
Creates a boxplot for LAI data
}
