% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createFolders}
\alias{createFolders}
\title{Creates (sub)folders for converted data and reports}
\usage{
createFolders(file, inputfolder, outputfolder)
}
\arguments{
\item{file}{filename}

\item{inputfolder}{input folder}

\item{outputfolder}{output folder}
}
\description{
Creates (sub)folders for converted data and reports
}
