% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createGridPlot}
\alias{createGridPlot}
\title{Plots the PlotNr arranged in a grid}
\usage{
createGridPlot(griddata)
}
\arguments{
\item{griddata}{grid data}
}
\description{
Plots the PlotNr arranged in a grid
}
