% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createGridPlotLAI}
\alias{createGridPlotLAI}
\title{Plots LAI on a grid}
\usage{
createGridPlotLAI(data, griddata)
}
\arguments{
\item{data}{imported sunscan data}

\item{griddata}{grid data}
}
\description{
Plots LAI on a grid
}
