% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createSeriesInfo}
\alias{createSeriesInfo}
\title{Summarise information about measurement series}
\usage{
createSeriesInfo(data)
}
\arguments{
\item{data}{imported sunscan data}
}
\value{
dataframe
}
\description{
Summarise information about measurement series
}
