% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createSummary}
\alias{createSummary}
\title{Summarises data for each PlotID}
\usage{
createSummary(data)
}
\arguments{
\item{data}{data.frame with LAI data}
}
\value{
data.frame summary information
}
\description{
Summarises data for each PlotID
}
