% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{createTimePlot}
\alias{createTimePlot}
\title{Plots measurements over time}
\usage{
createTimePlot(data, stripes = 11)
}
\arguments{
\item{data}{imported sunscan data}

\item{stripes}{number of color stripes}
}
\description{
Plots measurements over time
}
