% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{generateInitialGridData}
\alias{generateInitialGridData}
\title{Creates initial grid data}
\usage{
generateInitialGridData(data, rows = 1, rowwise = FALSE)
}
\arguments{
\item{data}{converted sunscan data}

\item{rows}{number of rows of the field}

\item{rowwise}{TRUE if the numbering is rowwise}
}
\value{
data frame with grids (col/row)
}
\description{
Creates initial grid data
}
