% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{generateReport}
\alias{generateReport}
\title{Generates report concerning converted data}
\usage{
generateReport(file, inputfolder, outputfolder, reportscript)
}
\arguments{
\item{file}{filename}

\item{inputfolder}{input folder}

\item{outputfolder}{output folder}

\item{reportscript}{script to generate the report}
}
\description{
Generates report concerning converted data
}
