% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{generateSampleMeasurementIdData}
\alias{generateSampleMeasurementIdData}
\title{Creates Sample MeasurementID data}
\usage{
generateSampleMeasurementIdData(data, interval = 0)
}
\arguments{
\item{data}{converted sunscan data}

\item{interval}{time interval after which a new PlotNR is assumed}
}
\value{
data frame with ids
}
\description{
Creates Sample MeasurementID data
}
