% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{generateSamplePlotIdData}
\alias{generateSamplePlotIdData}
\alias{generateInitialPlotIdData}
\alias{generateInitialMeasurementIdData}
\title{Generates sample ID data}
\usage{
generateSamplePlotIdData(data)

generateInitialPlotIdData(data)

generateInitialMeasurementIdData(data)
}
\arguments{
\item{data}{LAI-Data from SunscanFile}
}
\value{
dataframe with IDs
}
\description{
Generates sample data to relate PlotNr
to grids, PlotIDs or MeasurementIDs
}
