% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{getData}
\alias{getData}
\title{Extracts LAI data from the file lines}
\usage{
getData(lines, date)
}
\arguments{
\item{lines}{vector of lines read from the data file}

\item{date}{start date of measurements from meta data}
}
\value{
data.frame with LAI measurements
}
\description{
Extracts LAI data from the file lines
}
