% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{getFileList}
\alias{getFileList}
\title{Fetches recursively all possible data files from a directory}
\usage{
getFileList(
  directory,
  prefix = "",
  extension = ".TXT",
  excludedir = "converted"
)
}
\arguments{
\item{directory}{to search for data files}

\item{prefix}{of files to take into account}

\item{extension}{of files to take into account}

\item{excludedir}{subdirectory name to exclude}
}
\value{
vector of potential data files
}
\description{
Fetches recursively all possible data files from a directory
}
