% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{getHeader}
\alias{getHeader}
\title{Gets meta data from the file header}
\usage{
getHeader(lines, path)
}
\arguments{
\item{lines}{vector of lines read from the data file}

\item{path}{path of the file}
}
\value{
named vector with meta data
}
\description{
Gets meta data from the file header
}
