% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{getParData}
\alias{getParData}
\title{Extracts PAR data from the file lines}
\usage{
getParData(lines, date)
}
\arguments{
\item{lines}{vector of lines read from the data file}

\item{date}{start date of measurements from meta data}
}
\value{
data.frame with PAR measurements
}
\description{
Extracts PAR data from the file lines
}
