% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{getSmallHeader}
\alias{getSmallHeader}
\title{Gets meta data from the file header}
\usage{
getSmallHeader(lines)
}
\arguments{
\item{lines}{vector of lines read from the data file}
}
\value{
named vector with meta data
}
\description{
Gets meta data from the file header
}
