% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{headerFileName}
\alias{headerFileName}
\alias{dataFileName}
\alias{pardataFileName}
\alias{summaryFileName}
\alias{boxplotFileName}
\alias{reportFileName}
\title{File name for the metadata file}
\usage{
headerFileName(file, outputfolder)

dataFileName(file, outputfolder)

pardataFileName(file, outputfolder)

summaryFileName(file, outputfolder)

boxplotFileName(file, outputfolder)

reportFileName(file, outputfolder)
}
\arguments{
\item{file}{filename of data file}

\item{outputfolder}{output folder}
}
\value{
filename
}
\description{
File name for the metadata file

File name for the LAI data file

File name for the PAR data file

File name for the summary file

File name for the boxplot image

File name for the report
}
