% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{isDataLine}
\alias{isDataLine}
\title{Checks if the line is a data line}
\usage{
isDataLine(line)
}
\arguments{
\item{line}{a line of text}
}
\value{
TRUE if data line
}
\description{
Data lines begin with a date
}
