% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{isMovedFile}
\alias{isMovedFile}
\title{Checks if the data file has been moved to subfolder 'original'}
\usage{
isMovedFile(file, inputfolder, target = "original")
}
\arguments{
\item{file}{filename}

\item{inputfolder}{input folder}

\item{target}{target folder}
}
\value{
TRUE if file has been moved to original folder
}
\description{
Checks if the data file has been moved to subfolder 'original'
}
