% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{isSunscanFile}
\alias{isSunscanFile}
\title{Checks if given file is a valid SunScan file}
\usage{
isSunscanFile(path)
}
\arguments{
\item{path}{path to the file}
}
\value{
TRUE if valid SunScan file
}
\description{
Checks if file starts with 'Created by SunData'
and if the files has at least 14 lines
}
