% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{isTitleLine}
\alias{isTitleLine}
\title{Checks if the line is a title line}
\usage{
isTitleLine(line)
}
\arguments{
\item{line}{a line of text}
}
\value{
TRUE if title line
}
\description{
A title line starts a new metatdata section in the
sunscan data file.
}
