% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{movedFileName}
\alias{movedFileName}
\title{Returns path for the data file moved to 'original' subfolder}
\usage{
movedFileName(file, inputfolder, target = "original")
}
\arguments{
\item{file}{filename}

\item{inputfolder}{input folder}

\item{target}{target folder}
}
\value{
path to data file
}
\description{
Returns path for the data file moved to 'original' subfolder
}
