% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{moveFilesToSubfolders}
\alias{moveFilesToSubfolders}
\title{Moves data file to subfolder 'original'}
\usage{
moveFilesToSubfolders(file, inputfolder, target = "original")
}
\arguments{
\item{file}{filename}

\item{inputfolder}{input folder}

\item{target}{target folder}
}
\description{
Moves data file to subfolder 'original'
}
