% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{readConvertedHeader}
\alias{readConvertedHeader}
\alias{readConvertedData}
\alias{readConvertedParData}
\alias{readConvertedSummary}
\alias{readConvertedFiles}
\title{Reads converted data}
\usage{
readConvertedHeader(file, outputfolder)

readConvertedData(file, outputfolder)

readConvertedParData(file, outputfolder)

readConvertedSummary(file, outputfolder)

readConvertedFiles(file, outputfolder)
}
\arguments{
\item{file}{filename}

\item{outputfolder}{output folder}
}
\value{
dataframe with converted data
}
\description{
Reads converted data
}
