% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{readIdData}
\alias{readIdData}
\title{Read dataframe with PlotID column}
\usage{
readIdData(file)
}
\arguments{
\item{file}{filename}
}
\value{
data frame with the ids
}
\description{
Read dataframe with PlotID column
}
