% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{reorderDataColumns}
\alias{reorderDataColumns}
\title{Reorders columns in converted data}
\usage{
reorderDataColumns(data)
}
\arguments{
\item{data}{imported sunscan data}
}
\value{
dataframe with reordered columns
}
\description{
Reorders columns in converted data
}
