% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{saveHeader}
\alias{saveHeader}
\alias{saveData}
\alias{saveParData}
\alias{saveSummary}
\alias{saveBoxplot}
\title{Saves metadata to file}
\usage{
saveHeader(data, file, outputfolder)

saveData(data, file, outputfolder)

saveParData(data, file, outputfolder)

saveSummary(data, file, outputfolder)

saveBoxplot(data, file, outputfolder)
}
\arguments{
\item{data}{data frame}

\item{file}{filename of original file}

\item{outputfolder}{output folder}
}
\description{
Saves metadata to file

Saves LAI data

Saves PAR data

Saves summary

Saves boxplot
}
