% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{splitLines}
\alias{splitLines}
\title{Splits the data when file contains multiple headers}
\usage{
splitLines(lines)
}
\arguments{
\item{lines}{vector of lines from the data file}
}
\value{
list of line vectors
}
\description{
Splits the data when file contains multiple headers
}
