% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{stripFileExtension}
\alias{stripFileExtension}
\title{Strips the file extension}
\usage{
stripFileExtension(file)
}
\arguments{
\item{file}{filename}
}
\value{
file name without path and extension
}
\description{
Strips the file extension
}
