% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\docType{package}
\name{sunscanimport}
\alias{sunscanimport}
\title{Functions to import, convert and visualise data from LAI Sunscan device}
\description{
Functions to:
\itemize{
\item Convert a file or a directory
\item Create reports for converted data
\item run a shiny app for interactive conversion
}
}
\examples{
\dontrun{
  library(sunscanimport)
  runSunscanApp()
}

\dontrun{
  library(sunscanimport)
  file <- "paulinenaue.TXT"
  inputfolder <-  "data/210908/original/"
  outputfolder <- "data/210908/"
  convfile <- convertSunscanFile(file,inputfolder, outputfolder)
  generateReport(convfile, inputfolder, outputfolder)
}

\dontrun{
  library(sunscanimport)
  convertSunScanDirectory("data/")
}

}
\author{
{Gunther Krauss}
}
