% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{transformToWideFormat}
\alias{transformToWideFormat}
\title{Transform LAI data to wide format (1 row per PlotID)}
\usage{
transformToWideFormat(data, deleted = FALSE)
}
\arguments{
\item{data}{data.frame with LAI data}

\item{deleted}{include measurements marked as deleted}
}
\value{
data.frame in wide format
}
\description{
Each PlotID has only one row where each individual LAI
measurements is in an own column. Number of columns is determined by
the maxiumum number of measurements per PlotID. Measurements are populated
from left to right and filled up by NAs.
}
