% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\name{generateReport}
\alias{generateReport}
\title{Generates report concerning converted data}
\usage{
generateReport(
  file,
  inputfolder,
  outputfolder,
  reportscript = system.file("reports", "default_report.R", package = "sunscanimport")
)
}
\arguments{
\item{file}{filename}

\item{inputfolder}{input folder}

\item{outputfolder}{output folder}

\item{reportscript}{script to generate the report}
}
\description{
If one does not provide a report script, then the default report script is
used.
}
\details{
A report script is an R script that may contain RMarkdown. Inside the script
one can use the variable \code{param} with its elements \code{param$file}, \code{param$inputfolder} and \code{param$outputfolder}.
The report will be placed in the outputfolder in the subdirectory \code{report}.

Notice that the package \code{rmarkdown} and \code{knitr} have to be installed to generate a report.
}
