% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunscan_functions.R
\docType{package}
\name{sunscanimport}
\alias{sunscanimport}
\title{Functions to import, convert and visualise data from LAI Sunscan device}
\description{
The package provides functions to:
\itemize{
\item convert a file or a directory
\item summarize or transform converted data
\item create reports for converted data
\item run a shiny app for interactive conversion
}
}
\examples{
\dontrun{
  runSunscanApp()}

\dontrun{
  file <- "paulinenaue.TXT"
  inputfolder <-  "data/210908/original/"
  outputfolder <- "data/210908/"
  convfile <- convertSunscanFile(file,inputfolder, outputfolder)
  generateReport(convfile, inputfolder, outputfolder)}

\dontrun{
  convertSunScanDirectory("data/")}

\dontrun{
  data <- readr::read_delim("data/210908/converted/data_paulinenaue.txt", delim="\t")
  data_summary <- createSummary(data)
  data_wide <- transformToWideFormat(data)}

}
\author{
{Gunther Krauss}
}
